feat(render+ui): view-relative GPU tessellation + double-precision cursor, pan, snap, preview
Wires the viewport, cursor, snap, and rubber/transform preview pipelines
through PR #N (coordinate frame) so high-zoom and large-coordinate drawings
stop showing warped circles, steppy cursor, and stale preview tessellation.

ViewportRenderer:
- Pan stored / threaded as double end-to-end; geometry emitted in
  view-relative coordinates (double subtract -> float on GPU).
- Grid pan-anchored in view space.
- Committed-geometry GPU cache now keyed on (revision, view anchor, halfH,
  fb height). At high zoom the circle segment count is recomputed instead
  of replaying stale low-segment caches.

CadUi / main:
- DrawDrawingViewport signature: double* pan, double* cursor, double*
  outCursorRaw, plus a CadSnap::Hit out-param.
- Wheel zoom and middle-button pan in double.
- Cursor / snap / grip-magnet path is double end-to-end (no float
  quantization).
- Status bar consumes double cursor.

CadSnap:
- FindBest signature accepts double world coords so high-zoom picks don't
  step in pixels.

CadRubberPreview / TransformPreview:
- AppendCadDraftRubberLines and PushRubberSegViewRel receive the view
  anchor + orthoHalfH + fbHeight so rubber circles get the same adaptive
  segment count as committed circles.
- World-space output preserved (renderer does the precision-safe
  world->view conversion once).

+++ src/ui/CadUi.hpp
@@ -33,15 +33,15 @@ void DrawCommandLinePanel(std::vector<std::string>& log, char* cmdBuf, int cmdBu
 /// Fixed-height strip: OSNAP, ORTHO, GRID, POLAR, plot scale, cursor readout. Laid out by \ref main.cpp across the
 /// bottom of the main viewport (not docked, not movable).
 float CadStatusBarStripHeightPx();
-void DrawCadStatusBarStrip(AppCommandState& cmd, float cursorX, float cursorY, float cursorZ,
+void DrawCadStatusBarStrip(AppCommandState& cmd, double cursorX, double cursorY, float cursorZ,
                            bool* ortho_mode_enabled, bool* grid_visible);
 
 /// Central CAD viewport: renders OpenGL texture and handles pan / zoom / LINE picks.
 /// Writes framebuffer pixel size and cursor world position. When object snap finds a hit, cursor and
 /// \p out_snap reflect the snapped point; otherwise raw hover coordinates.
 void DrawDrawingViewport(unsigned int viewportTextureId, AppCommandState& cmd, std::vector<std::string>& log,
-                         char* cmdBuf, int cmdBufSize, float* panX, float* panY, float* zoom, float* outCursorX,
-                         float* outCursorY, float* outCursorRawX, float* outCursorRawY, int* outFbW, int* outFbH,
+                         char* cmdBuf, int cmdBufSize, double* panX, double* panY, float* zoom, double* outCursorX,
+                         double* outCursorY, double* outCursorRawX, double* outCursorRawY, int* outFbW, int* outFbH,
                          CadSnap::Hit* out_snap);
 
 void DrawCreatePointsPanel(AppCommandState& cmd, std::vector<std::string>& log);
